Adopt typed throws across 'BinaryParsing'
To accommodate both user and library errors on parser function-taking
functions (like Range.init(parsingStartAndEnd:parser:)), this adds
a 'typealias ThrownParsingError' that aliases 'ParsingError' in
embedded mode and 'any Error' elsewhere. This has the benefit of
keeping the simplicity of untyped errors in non-embedded code, but
means that adopting code can't choose to use explicitly typed errors
when in non-embedded mode.