Revise the typed throws design
This changes the conventions for typed throws for the parser
initializers that used the conditionally-defined `ThrownParsingError`
type alias. Some of the parsers didn't actually need that alias,
and instead could use a function-specific generic error.

For others parsers, the library now provides two overloads: one with
`ParsingError`-typed errors and one with untyped errors, which is
omitted in embedded contexts. Since it is more specific, the typed-
error overload is chosen when a passed-in closure either does not
throw or throws only `ParsingError`s.