Restrict relative seeking a bit further
This change means that the `seek(toRelativeOffset:)` and
`seek(toOffsetFromEnd:)` are restricted to movement _within_
the current range of the span. This is how these are used, and is
a bit easier to explain because movement is therefore bound by both
ends of the span.

The `seek(toAbsoluteOffset:)` method still works with the original,
un-shrunken bounds, so has a different set of semantics (and may
benefit from a different name).